# Conflicts:
#	app/Http/Controllers/ExportController.php
#	public/js/itemList.js
#	resources/views/character/loot.blade.php
#	routes/web.php

+++ app/Http/Controllers/AssignLootController.php
@@ -58,7 +58,7 @@ class AssignLootController extends Controller
         $currentMember = request()->get('currentMember');
 
         if (!$currentMember->hasPermission('edit.raid-loot')) {
-            request()->session()->flash('status', 'You don\'t have permissions to view that page.');
+            request()->session()->flash('status', __("You don't have permissions to view that page."));
             return redirect()->route('member.show', ['guildId' => $guild->id, 'guildSlug' => $guild->slug, 'memberId' => $currentMember->id, 'usernameSlug' => $currentMember->slug]);
         }
 
@@ -97,7 +97,7 @@ class AssignLootController extends Controller
         $currentMember = request()->get('currentMember');
 
         if (!$currentMember->hasPermission('edit.raid-loot')) {
-            request()->session()->flash('status', 'You don\'t have permissions to view that page.');
+            request()->session()->flash('status', __("You don't have permissions to view that page."));
             return redirect()->route('member.show', ['guildId' => $guild->id, 'guildSlug' => $guild->slug, 'memberId' => $currentMember->id, 'usernameSlug' => $currentMember->slug]);
         }
 
@@ -147,7 +147,7 @@ class AssignLootController extends Controller
         $currentMember = request()->get('currentMember');
 
         if (!$currentMember->hasPermission('edit.raid-loot')) {
-            request()->session()->flash('status', 'You don\'t have permissions to submit that.');
+            request()->session()->flash('status', __("You don't have permissions to submit that."));
             return redirect()->route('member.show', ['guildId' => $guild->id, 'guildSlug' => $guild->slug, 'memberId' => $currentMember->id, 'usernameSlug' => $currentMember->slug]);
         }
 
@@ -231,7 +231,7 @@ class AssignLootController extends Controller
             'created_at'    => getDateTime(),
         ]);
 
-        request()->session()->flash('status', "Successfully updated loot assignment (containing {$itemCount} items).");
+        request()->session()->flash('status', __("Successfully updated loot assignment (containing :itemCount items).", ['itemCount' => $itemCount]));
 
         return redirect()->route('item.assignLoot.list', [
             'guildId'   => $guild->id,
@@ -357,7 +357,7 @@ class AssignLootController extends Controller
         $currentMember = request()->get('currentMember');
 
         if (!$currentMember->hasPermission('edit.raid-loot')) {
-            request()->session()->flash('status', 'You don\'t have permissions to submit that.');
+            request()->session()->flash('status', __("You don't have permissions to submit that."));
             return redirect()->route('member.show', ['guildId' => $guild->id, 'guildSlug' => $guild->slug, 'memberId' => $currentMember->id, 'usernameSlug' => $currentMember->slug]);
         }
 
@@ -491,11 +491,11 @@ class AssignLootController extends Controller
                             'created_at'    => $now,
                         ];
                     } else {
-                        $warnings .= (isset($item['label']) ? $item['label'] : $item['id']) . ' to character ID ' . $item['character_id'] . ', ';
+                        $warnings .= __(':item to character ID :character, ', ['item' => (isset($item['label']) ? $item['label'] : $item['id']), 'character' => $item['character_id']]);
                         $failedCount++;
                     }
                 } else {
-                    $warnings .= (isset($item['label']) ? $item['label'] : $item['id']) . ' to missing character, ';
+                    $warnings .= __(':item to missing character, ', ['item' => (isset($item['label']) ? $item['label'] : $item['id'])]);
                     $failedCount++;
                 }
             }
@@ -537,8 +537,8 @@ class AssignLootController extends Controller
                 $whereClause['character_items.is_received'] = 0;
             }
 
-            // Find wishlist for this item
-            $wishlistRow = CharacterItem::
+            // Find wishlist rows for this item
+            $wishlistRows = CharacterItem::
                 select('character_items.*')
                 // Look for both the original item and the possible token reward for the item
                 ->join('items', function ($join) {
@@ -547,45 +547,48 @@ class AssignLootController extends Controller
                 })
                 ->where($whereClause)
                 ->whereRaw("(items.item_id = {$detachRow['item_id']} OR items.parent_item_id = {$detachRow['item_id']})")
-                ->limit(1)
+                ->groupBy(['character_items.list_number', 'character_items.item_id'])
                 ->orderBy('character_items.is_received')
+                ->orderBy('character_items.list_number')
                 ->orderBy('character_items.order')
-                ->first();
+                ->get();
 
-            if ($wishlistRow) {
-                if ($deleteWishlist) {
-                    // Delete the one we found
-                    CharacterItem::where(['id' => $wishlistRow->id])->delete();
-                    $audits[] = [
-                        'description'   => 'System removed 1 wishlist item after character was assigned item',
-                        'type'          => Item::TYPE_WISHLIST,
-                        'member_id'     => $currentMember->id,
-                        'character_id'  => $wishlistRow->character_id,
-                        'guild_id'      => $currentMember->guild_id,
-                        'raid_group_id' => $wishlistRow->raid_group_id,
-                        'raid_id'       => $raidId,
-                        'item_id'       => $wishlistRow->item_id,
-                        'created_at'    => $now,
-                    ];
-                } else {
-                    CharacterItem::
-                        where(['id' => $wishlistRow->id])
-                        ->update([
-                            'is_received' => 1,
-                            'received_at' => $now,
-                        ]);
+            if ($wishlistRows->count()) {
+                foreach ($wishlistRows as $wishlistRow) {
+                    if ($deleteWishlist && $wishlistRow->list_number == $guild->current_wishlist_number) {
+                        // Delete the one we found
+                        CharacterItem::where(['id' => $wishlistRow->id])->delete();
+                        $audits[] = [
+                            'description'   => 'System removed 1 wishlist item after character was assigned item',
+                            'type'          => Item::TYPE_WISHLIST,
+                            'member_id'     => $currentMember->id,
+                            'character_id'  => $wishlistRow->character_id,
+                            'guild_id'      => $currentMember->guild_id,
+                            'raid_group_id' => $wishlistRow->raid_group_id,
+                            'raid_id'       => $raidId,
+                            'item_id'       => $wishlistRow->item_id,
+                            'created_at'    => $now,
+                        ];
+                    } else {
+                        CharacterItem::
+                            where(['id' => $wishlistRow->id])
+                            ->update([
+                                'is_received' => 1,
+                                'received_at' => $now,
+                            ]);
 
-                    $audits[] = [
-                        'description'   => 'System flagged 1 wishlist item as received after character was assigned item',
-                        'type'          => Item::TYPE_WISHLIST,
-                        'member_id'     => $currentMember->id,
-                        'character_id'  => $wishlistRow->character_id,
-                        'guild_id'      => $currentMember->guild_id,
-                        'raid_group_id' => $wishlistRow->raid_group_id,
-                        'raid_id'       => $raidId,
-                        'item_id'       => $wishlistRow->item_id,
-                        'created_at'    => $now,
-                    ];
+                        $audits[] = [
+                            'description'   => 'System flagged 1 wishlist item as received after character was assigned item',
+                            'type'          => Item::TYPE_WISHLIST,
+                            'member_id'     => $currentMember->id,
+                            'character_id'  => $wishlistRow->character_id,
+                            'guild_id'      => $currentMember->guild_id,
+                            'raid_group_id' => $wishlistRow->raid_group_id,
+                            'raid_id'       => $raidId,
+                            'item_id'       => $wishlistRow->item_id,
+                            'created_at'    => $now,
+                        ];
+                    }
                 }
             }
 
@@ -650,7 +653,15 @@ class AssignLootController extends Controller
 
         AuditLog::insert($audits);
 
-        request()->session()->flash('status', 'Successfully added ' . $addedCount . ' items. ' . $failedCount . ' failures' . ($warnings ? ': ' . rtrim($warnings, ', ') : '.'));
+        request()->session()->flash('status',
+            __('Successfully added :addedCount items. :failedCount failures:warnings',
+                [
+                    'addedCount'  => $addedCount,
+                    'failedCount' => $failedCount,
+                    'warnings'    => ($warnings ? ': ' . rtrim($warnings, ', ') : '.'),
+                ]
+            )
+        );
 
         return redirect()->route('guild.auditLog', [
             'guildId'   => $guild->id,
